requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
pyserial==3.5
flask==3.0.0
skyfield==1.46
numpy==1.24.3
